Fix brew install/upgrade function
Fixes function for installing or upgrading a brew formula by removing a
placeholder test and adding actual tests for install status.

It is possible that this revised function makes checking whether a given
formula is in Homebrew redundant in certain circumstances. Refactoring
or eliminating that test in the future may speed up munkitap.