Fix package symbolic link install error
On macOS, the directory /usr/local/bin/ is not created by default.
Therefore, the postinstall script that generated the munkitap symbolic
link would fail on systems where it did not already exist. This error
can be avoided by installing the symbolic link as part of the package
and not by a script. The pkg build script has been updated to
automatically include the symbolic link and the postinstall script has
been removed.

The package version has been bumped up to reflect this change.