Implement munkitap logging and quiet/verbose flags
Implements logging with munkitap using the MunkitapLog object. Output is
logged to a file in the munkitap cache, regardless of whether the
--quiet flag is passed (which only affects printing to the terminal). If
the --verbose flag is passed, the information printed to the terminal
and to the log file will be more verbose, and may be more useful for
debugging.

For future development, consider implementing JSON logs to be more
easily machine-readable and implementing more useful logging messages.